OctreeNode uses single precision to reduce memory use
git-svn-id: svn://svn.code.sf.net/p/aoi/code@409 8e91ba97-c92a-0410-817f-f646a3b87f38

+++ Renderers/src/artofillusion/raytracer/Raytracer.java
@@ -518,7 +518,7 @@ public class Raytracer
     
     // Create the octree.
 
-    rootNode = new OctreeNode(minx, maxx, miny, maxy, minz, maxz, sceneObject, objBounds, null);
+    rootNode = new OctreeNode((float) minx, (float) maxx, (float) miny, (float) maxy, (float) minz, (float) maxz, sceneObject, objBounds, null);
 
     // Find the nodes which contain the camera and the lights.
